Cleanup BucketMetadata public interface. (#1877)
Move ParseFromString(), ParseFromJson() and ToJson() to the internal
namespace. None of these functions should be called by application
developers, and therefore we do not want them exposed in the public API.
