+++ google/cloud/storage/internal/bucket_requests.h
@@ -27,6 +27,18 @@ namespace cloud {
 namespace storage {
 inline namespace STORAGE_CLIENT_NS {
 namespace internal {
+struct LifecycleRuleParser {
+  static StatusOr<LifecycleRule> FromJson(internal::nl::json const& json);
+  static StatusOr<LifecycleRule> FromString(std::string const& text);
+};
+
+struct BucketMetadataParser {
+  static StatusOr<BucketMetadata> FromJson(internal::nl::json const& json);
+  static StatusOr<BucketMetadata> FromString(std::string const& payload);
+};
+
+std::string BucketMetadataToJsonString(BucketMetadata const& meta);
+
 /**
  * Requests the list of buckets for a project.
  */
@@ -94,7 +106,9 @@ class CreateBucketRequest
       : project_id_(std::move(project_id)), metadata_(std::move(metadata)) {}
 
   /// Returns the request as the JSON API payload.
-  std::string json_payload() const { return metadata_.ToJsonString(); }
+  std::string json_payload() const {
+    return BucketMetadataToJsonString(metadata_);
+  }
 
   std::string const& project_id() const { return project_id_; }
   CreateBucketRequest& set_project_id(std::string project_id) {
@@ -147,7 +161,9 @@ class UpdateBucketRequest
       : metadata_(std::move(metadata)) {}
 
   /// Returns the request as the JSON API payload.
-  std::string json_payload() const { return metadata_.ToJsonString(); }
+  std::string json_payload() const {
+    return BucketMetadataToJsonString(metadata_);
+  }
 
   BucketMetadata const& metadata() const { return metadata_; }
